fav-task-changes

+++ src/modules/product/product.ts
@@ -11,38 +11,73 @@ export class Product {
   view: View;
   product: ProductData;
   params: ProductComponentParams;
-  removeButton: boolean;
 
-  constructor(product: ProductData, params: ProductComponentParams = {}, removeButton: boolean = false) {
+  constructor(product: ProductData, params: ProductComponentParams = {}) {
     this.product = product;
     this.params = params;
     this.view = new ViewTemplate(html).cloneView();
-    this.removeButton = removeButton;
   }
 
   attach($root: HTMLElement) {
     $root.appendChild(this.view.root);
   }
 
-  render() {
+  detach(e: MouseEvent) {
+    e.preventDefault();
+    favoritesService.removeProductFromFav(this.product);
+    this.view.root.remove();
+  }
+
+  async render() {
     const { id, name, src, salePriceU } = this.product;
 
     this.view.root.setAttribute('href', `/product?id=${id}`);
     this.view.img.setAttribute('src', src);
     this.view.title.innerText = name;
     this.view.price.innerText = formatPrice(salePriceU);
-    this.removeButton && this.view.removeButton.classList.add('removeButton')
     
-    this.view.removeButton.onclick = (event: Event) => {
-        event.preventDefault();
-        favoritesService.removeProductFromFav(this.product);
-        this.view.root.remove();
-    };
+    this.view.removeButton.onclick = (e: MouseEvent) => this.detach(e);
+
+    const isInFavorites = await favoritesService.isInFavorites(this.product);
     
-    if (this.params.isHorizontal) {
-        this.view.root.classList.add('is__horizontal')
-    } else {
-        this.view.root.classList.add('is__vertical')
-    }
+    this.view.btnToFavorites.onclick = async (e: MouseEvent) => {
+        e.preventDefault();
+        if (!this.product) return;
+    
+        const isInFavorites = await favoritesService.isInFavorites(this.product);
+        if (isInFavorites) {
+          favoritesService.removeProductFromFav(this.product);
+          this.view.btnToFavorites.classList.remove('inFavorites');
+        }
+        if (!isInFavorites) {
+          favoritesService.addProductToFav(this.product);
+          this.view.btnToFavorites.classList.add('inFavorites');
+        }
+      };
+
+    if (isInFavorites) this.view.btnToFavorites.classList.add('inFavorites');
+    
+    this.params.isHorizontal ? this.view.root.classList.add('is__horizontal') : this.view.root.classList.add('is__vertical');
+
+    if(window.location.pathname === '/favorites') this.view.removeButton.classList.add('is-shown')
+    if(window.location.pathname === '/' || window.location.pathname === '/catalog') this.view.btnToFavorites.classList.add('is-heart-shown');
+  }
+
+  async addToFavorites(e: MouseEvent) {
+    e.preventDefault();
+    if (!this.product) return;
+    if (await favoritesService.isInFavorites(this.product)) return;
+  
+    favoritesService.addProductToFav(this.product);
+    this.view.btnToFavorites.classList.add('inFavorites');
+  }
+  
+  async removeFromFavorites(e: MouseEvent) {
+    e.preventDefault();
+    if (!this.product) return;
+    if (!await favoritesService.isInFavorites(this.product)) return;
+  
+    favoritesService.removeProductFromFav(this.product);
+    this.view.btnToFavorites.classList.remove('inFavorites');
   }
 }
